Fixed documentation builder warnings
